adds retrospective TDD for Instrument subclasses play method with a few minor adjustments to var types

+++ src/main/java/instruments/Trombone.java
@@ -5,10 +5,10 @@ import behaviours.IPlay;
 public class Trombone extends Instrument implements IPlay {
 
     private String bodyDesign;
-    private int bellDiameter;
-    private int boreDiameter;
+    private double bellDiameter;
+    private double boreDiameter;
 
-    public Trombone(String brand, String model, String type, double cost, double retail, String bodyDesign, int bellDiameter, int boreDiameter) {
+    public Trombone(String brand, String model, String type, double cost, double retail, String bodyDesign, double bellDiameter, double boreDiameter) {
         super(brand, model, type, cost, retail);
         this.bodyDesign = bodyDesign;
         this.bellDiameter = bellDiameter;
@@ -19,11 +19,11 @@ public class Trombone extends Instrument implements IPlay {
         return bodyDesign;
     }
 
-    public int getBellDiameter() {
+    public double getBellDiameter() {
         return bellDiameter;
     }
 
-    public int getBoreDiameter() {
+    public double getBoreDiameter() {
         return boreDiameter;
     }
 
@@ -31,11 +31,11 @@ public class Trombone extends Instrument implements IPlay {
         this.bodyDesign = bodyDesign;
     }
 
-    public void setBellDiameter(int bellDiameter) {
+    public void setBellDiameter(double bellDiameter) {
         this.bellDiameter = bellDiameter;
     }
 
-    public void setBoreDiameter(int boreDiameter) {
+    public void setBoreDiameter(double boreDiameter) {
         this.boreDiameter = boreDiameter;
     }
